+++ CHANGES.md
@@ -54,6 +54,7 @@
 
 ## v0.7.1
 ```
+- Record packet length properly.
 - Remove dependencies related to READING_FIRMWARE_PREALLOC_BUFFER (which was removed from mainline).
 - Made git dependencies explicit.
 - Rename build (SECURITY_PROVENANCE_WHOLE_SYSTEM -> SECURITY_PROVENANCE_BOOT).